fix: adapting the structure to merge
