+++ app/src/main/java/com/android/sample/model/CreatureStats.kt
@@ -1,0 +1,8 @@
+package com.android.sample.model
+
+data class CreatureStats(
+    val happiness: Int = 85,
+    val health: Int = 90,
+    val energy: Int = 70,
+    val level: Int = 5,
+)
