+++ app/src/main/java/com/android/sample/model/UserStats.kt
@@ -1,0 +1,8 @@
+package com.android.sample.model
+
+data class UserStats(
+    val streakDays: Int = 7,
+    val points: Int = 1250,
+    val studyTodayMin: Int = 45,
+    val dailyGoalMin: Int = 180,
+)
